feat: Event 목록을 반환하는 DTO 객체를 별도로 관리하도록 추가

+++ src/main/java/kr/co/dbsg/api/api/event/service/EventService.java
@@ -24,7 +24,7 @@ public class EventService {
     }
 
     @Transactional(readOnly = true)
-    public Event getEvent(int id) {
+    public Event getEvent(long id) {
         return eventRepository.findById(id)
             .orElseThrow(IllegalArgumentException::new)
             .toEvent();
